Tidy up util module
Rather than have a catch-all module, split it up. Added some more
rustdocs, and changed the signature and tests for CPU range parsing to
not do any io anymore.

Test Plan
=========

CI
