+++ src/util/mod.rs
@@ -1,0 +1,6 @@
+mod cpu;
+mod lpm;
+
+pub use cpu::get_online_cpus;
+pub use lpm::summarize_address_range;
+pub use lpm::AddressBlockRange;
